dj-app/requirements.txt
alabaster==0.7.3
appdirs==1.4.2
Babel==1.3
coverage==4.0a5
dj-database-url==0.4.1
Django==1.10
django-bootstrap-form==3.2
django-bootstrap-forms==0.1
django-bower==5.0.2
django-cors-headers==2.0.2
django-nose==1.3
django-tastypie==0.13.3
djangorestframework==3.5.4
djangorestframework-jwt==1.10.0
djoser==0.5.4
docutils==0.12
gunicorn==19.6.0
httplib2==0.9
httpretty==0.8.8
Jinja2==2.7.3
MarkupSafe==0.23
nose==1.3.4
oauth2==1.5.211
oauthlib==0.7.2
packaging==16.8
psycopg2==2.6.2
Pygments==2.0.2
PyJWT==1.4.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2016.6.1
requests==2.6.0
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9